# opt: reduce allocations in Histogram.DistinctValuesCount