@mgartner 2022-12-13
*[`pkg/sql/opt/props/histogram.go` line 188 at r4](https://reviewable.io/reviews/cockroachdb/cockroach/93482#-NJ9Yh0p3iYTJhFOBAf5:-NJB4Su_EGDEdiCoflts:b-g4c4qk) ([raw file](https://github.com/cockroachdb/cockroach/blob/aa832e10a9ef82d891344c242211d06e127b7b90/pkg/sql/opt/props/histogram.go#L188)):*
<details><summary><i>Previously, DrewKimball (Drew Kimball) wrote…</i></summary><blockquote>

Not directly related to this PR, but should we be using `tree.MaxDistinctCount` here?
</blockquote></details>

Ya we probably should. I've left a TODO and will look into in a follow-up PR if I get a chance.


<!-- Sent from Reviewable.io -->